test: 리뷰 반영 - Name 클래스 관련 테스트 코드 수정
