+++ src/test/java/carracing/ResultViewTest.java
@@ -2,6 +2,7 @@ package carracing;
 
 import carracing.exception.CarNameFormatException;
 import carracing.util.Car;
+import carracing.util.Name;
 import carracing.view.ResultView;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,7 +20,7 @@ public class ResultViewTest {
         List<Car> carList = new ArrayList<>();
 
         for (int i = 0; i < 10; i++) {
-            carList.add(new Car(String.valueOf(i)));
+            carList.add(new Car(new Name(String.valueOf(i))));
         }
 
         resultView.printRacing(carList);
@@ -29,7 +30,7 @@ public class ResultViewTest {
     @DisplayName("자동차 이름 출력 확인")
     void printCarName() throws CarNameFormatException {
         ResultView resultView = new ResultView();
-        resultView.print(new Car("name"));
+        resultView.print(new Car(new Name("name")));
     }
 
     @Test
